+++ src/burn/engine/precomp.h
@@ -73,6 +73,7 @@
 #include "condition.h"
 #include "search.h"
 #include "section.h"
+#include "approvedexe.h"
 #include "container.h"
 #include "catalog.h"
 #include "payload.h"