+++ src/burn/inc/IBootstrapperEngine.h
@@ -217,4 +217,12 @@ DECLARE_INTERFACE_IID_(IBootstrapperEngine, IUnknown, "6480D616-27A0-44D7-905B-8
     STDMETHOD(Quit)(
         __in DWORD dwExitCode
         ) = 0;
+
+    STDMETHOD(LaunchApprovedExe)(
+        __in_opt HWND hwndParent,
+        __in_z LPCWSTR wzApprovedExeForElevationId,
+        __in_z LPCWSTR wzExecutablePath,
+        __in_z_opt LPCWSTR wzArguments,
+        __in DWORD dwWaitForInputIdleTimeout
+        ) = 0;
 };