+++ src/ext/BalExtension/mba/core/IBootstrapperEngine.cs
@@ -145,6 +145,14 @@ namespace Microsoft.Tools.WindowsInstallerXml.Bootstrapper
         void Quit(
             [MarshalAs(UnmanagedType.U4)] int dwExitCode
             );
+
+        void LaunchApprovedExe(
+            IntPtr hwndParent,
+            [MarshalAs(UnmanagedType.LPWStr)] string wzApprovedExeForElevationId,
+            [MarshalAs(UnmanagedType.LPWStr)] string wzExecutablePath,
+            [MarshalAs(UnmanagedType.LPWStr)] string wzArguments,
+            [MarshalAs(UnmanagedType.U4)] int dwWaitForInputIdleTimeout
+            );
     }
 
     /// <summary>