+++ src/ext/BalExtension/wixext/BalCompiler.cs
@@ -211,6 +211,9 @@ namespace Microsoft.Tools.WindowsInstallerXml.Extensions
         {
             SourceLineNumberCollection sourceLineNumbers = Preprocessor.GetSourceLineNumbers(node);
             string launchTarget = null;
+            string launchTargetElevatedId = null;
+            string launchArguments = null;
+            YesNoType launchHidden = YesNoType.NotSet;
             string licenseFile = null;
             string licenseUrl = null;
             string logoFile = null;
@@ -230,6 +233,15 @@ namespace Microsoft.Tools.WindowsInstallerXml.Extensions
                     {
                         case "LaunchTarget":
                             launchTarget = this.Core.GetAttributeValue(sourceLineNumbers, attrib, false);
+                            break;
+                        case "LaunchTargetElevatedId":
+                            launchTargetElevatedId = this.Core.GetAttributeIdentifierValue(sourceLineNumbers, attrib);
+                            break;
+                        case "LaunchArguments":
+                            launchArguments = this.Core.GetAttributeValue(sourceLineNumbers, attrib, false);
+                            break;
+                        case "LaunchHidden":
+                            launchHidden = this.Core.GetAttributeYesNoValue(sourceLineNumbers, attrib);
                             break;
                         case "LicenseFile":
                             licenseFile = this.Core.GetAttributeValue(sourceLineNumbers, attrib, false);
@@ -297,6 +309,21 @@ namespace Microsoft.Tools.WindowsInstallerXml.Extensions
                 if (!String.IsNullOrEmpty(launchTarget))
                 {
                     this.Core.CreateVariableRow(sourceLineNumbers, "LaunchTarget", launchTarget, "string", false, false);
+                }
+
+                if (!String.IsNullOrEmpty(launchTargetElevatedId))
+                {
+                    this.Core.CreateVariableRow(sourceLineNumbers, "LaunchTargetElevatedId", launchTargetElevatedId, "string", false, false);
+                }
+
+                if (!String.IsNullOrEmpty(launchArguments))
+                {
+                    this.Core.CreateVariableRow(sourceLineNumbers, "LaunchArguments", launchArguments, "string", false, false);
+                }
+
+                if (YesNoType.Yes == launchHidden)
+                {
+                    this.Core.CreateVariableRow(sourceLineNumbers, "LaunchHidden", "yes", "string", false, false);
                 }
 
                 if (!String.IsNullOrEmpty(licenseFile))