+++ src/tools/wix/Binder.cs
@@ -3640,6 +3640,19 @@ namespace Microsoft.Tools.WindowsInstallerXml
                 this.ResolveDelayedFields(bundle, delayedFields, variableCache, null);
             }
 
+            // Process WixApprovedExeForElevation rows.
+            Table wixApprovedExeForElevationTable = bundle.Tables["WixApprovedExeForElevation"];
+            List<ApprovedExeForElevation> approvedExesForElevation = new List<ApprovedExeForElevation>();
+            if (null != wixApprovedExeForElevationTable && 0 < wixApprovedExeForElevationTable.Rows.Count)
+            {
+                foreach (WixApprovedExeForElevationRow wixApprovedExeForElevationRow in wixApprovedExeForElevationTable.Rows)
+                {
+                    ApprovedExeForElevation approvedExeForElevation = new ApprovedExeForElevation(wixApprovedExeForElevationRow);
+                    approvedExeForElevation.CalculateHash();
+                    approvedExesForElevation.Add(approvedExeForElevation);
+                }
+            }            
+
             // Set the overridable bundle provider key.
             this.SetBundleProviderKey(bundle, bundleInfo);
 
@@ -3707,7 +3720,7 @@ namespace Microsoft.Tools.WindowsInstallerXml
             }
 
             string manifestPath = Path.Combine(this.TempFilesLocation, "bundle-manifest.xml");
-            this.CreateBurnManifest(bundleFile, bundleInfo, bundleUpdateRow, updateRegistrationInfo, manifestPath, allRelatedBundles, allVariables, orderedSearches, allPayloads, chain, containers, catalogs, bundle.Tables["WixBundleTag"]);
+            this.CreateBurnManifest(bundleFile, bundleInfo, bundleUpdateRow, updateRegistrationInfo, manifestPath, allRelatedBundles, allVariables, orderedSearches, allPayloads, chain, containers, catalogs, bundle.Tables["WixBundleTag"], approvedExesForElevation);
 
             this.UpdateBurnResources(bundleTempPath, bundleFile, bundleInfo);
 
@@ -4126,7 +4139,7 @@ namespace Microsoft.Tools.WindowsInstallerXml
             }
         }
 
-        private void CreateBurnManifest(string outputPath, WixBundleRow bundleInfo, WixBundleUpdateRow updateRow, WixUpdateRegistrationRow updateRegistrationInfo, string path, List<RelatedBundleInfo> allRelatedBundles, List<VariableInfo> allVariables, List<WixSearchInfo> orderedSearches, Dictionary<string, PayloadInfoRow> allPayloads, ChainInfo chain, Dictionary<string, ContainerInfo> containers, Dictionary<string, CatalogInfo> catalogs, Table wixBundleTagTable)
+        private void CreateBurnManifest(string outputPath, WixBundleRow bundleInfo, WixBundleUpdateRow updateRow, WixUpdateRegistrationRow updateRegistrationInfo, string path, List<RelatedBundleInfo> allRelatedBundles, List<VariableInfo> allVariables, List<WixSearchInfo> orderedSearches, Dictionary<string, PayloadInfoRow> allPayloads, ChainInfo chain, Dictionary<string, ContainerInfo> containers, Dictionary<string, CatalogInfo> catalogs, Table wixBundleTagTable, List<ApprovedExeForElevation> approvedExesForElevation)
         {
             string executableName = Path.GetFileName(outputPath);
 
@@ -4550,6 +4563,19 @@ namespace Microsoft.Tools.WindowsInstallerXml
                         writer.WriteStartElement("PatchTargetCode");
                         writer.WriteAttributeString("TargetCode", targetCode.TargetCode);
                         writer.WriteAttributeString("Product", targetCode.TargetsProductCode ? "yes" : "no");
+                        writer.WriteEndElement();
+                    }
+                }
+
+                // write the ApprovedExeForElevation elements
+                if (0 < approvedExesForElevation.Count)
+                {
+                    foreach (ApprovedExeForElevation approvedExeForElevation in approvedExesForElevation)
+                    {
+                        writer.WriteStartElement("ApprovedExeForElevation");
+                        writer.WriteAttributeString("Id", approvedExeForElevation.Id);
+                        writer.WriteAttributeString("FileSize", approvedExeForElevation.FileSize.ToString(CultureInfo.InvariantCulture));
+                        writer.WriteAttributeString("Hash", approvedExeForElevation.Hash);
                         writer.WriteEndElement();
                     }
                 }