+++ src/tools/wix/CompilerCore.cs
@@ -741,6 +741,17 @@ namespace Microsoft.Tools.WindowsInstallerXml
         }
 
         /// <summary>
+        /// Creates a WixApprovedExeForElevation row in the active session.
+        /// </summary>
+        /// <param name="sourceLineNumbers">Source and line number of current row.</param>
+        public void CreateWixApprovedExeForElevationRow(SourceLineNumberCollection sourceLineNumbers, string id, string sourceFile)
+        {
+            WixApprovedExeForElevationRow wixApprovedExeForElevationRow = (WixApprovedExeForElevationRow)this.CreateRow(sourceLineNumbers, "WixApprovedExeForElevation");
+            wixApprovedExeForElevationRow.Id = id;
+            wixApprovedExeForElevationRow.SourceFile = sourceFile;
+        }
+
+        /// <summary>
         /// Creates a WixCatalog row in the active section.
         /// </summary>
         /// <param name="sourceLineNumbers">Source and line number of current row.</param>