+++ test/src/UnitTests/Burn/precomp.h
@@ -51,6 +51,7 @@
 #include "condition.h"
 #include "search.h"
 #include "section.h"
+#include "approvedexe.h"
 #include "container.h"
 #include "catalog.h"
 #include "payload.h"